aula 04

+++ .Estudos blog prof. Piva/04_aula/exemplos/exemplo1_area-quadrilatero.py
@@ -1,6 +1,6 @@
 #Calcular a área de um quadrilátero, com dois lados iguais. Para isso faz a leitura de dois numeros int, representando os lados e imrpime os valores lidos e a área calculdada.
 
-altura = int(input("Digite o valor altura: "))
-base = int(input("Digite o valor da base: "))
+altura = float(input("Digite o valor altura: "))
+base = float(input("Digite o valor da base: "))
 area = base * altura
 print(f"A área é: {area}. Altura é: {altura}. A base é: {base}")
